Move MultiMatrix out of its own folder
